requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
starlette==0.27.0

# Authentication & Security
firebase-admin==6.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.0.1

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
sqlite3==0.0.0

# LangChain & AI
langchain==0.0.340
langchain-openai==0.0.2
langchain-community==0.0.3
openai==1.3.0
tiktoken==0.5.1

# RAG & Vector Stores
chromadb==0.4.15
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.35.0

# Document Processing
pypdf==3.17.0
python-docx==0.8.11
python-pptx==0.6.21
docx2txt==0.8

# Data Science & Analytics
pandas==2.1.0
numpy==1.25.0
scikit-learn==1.3.2
scipy==1.11.4
statsmodels==0.14.0

# Visualization
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2

# Streamlit & Web UI
streamlit==1.28.1
streamlit-authenticator==0.2.3
streamlit-option-menu==0.3.6

# HTTP Client & Utilities
requests==2.31.0
aiofiles==23.2.0
aiohttp==3.9.0
httpx==0.25.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Additional Utilities
pydantic==2.5.0
typing-extensions==4.8.0
redis==5.0.1
celery==5.3.4 